dns_random/requirements.txt
dnspython==2.7.0
dnslib==0.9.26
